HW8 commit 01.06 Реализованы продукты, заказы, корзина, юзеры. Работает через эврику и зул, частично покрыто тестами. Планируется полное покрытие тестами и доделка фронта

+++ router/src/main/java/ru/gb/market/happy/router/dto/OrderItemDto.java
@@ -2,10 +2,6 @@ package ru.gb.market.happy.router.dto;
 
 import lombok.Data;
 import lombok.NoArgsConstructor;
-import org.hibernate.annotations.CreationTimestamp;
-import org.hibernate.annotations.UpdateTimestamp;
-
-import javax.persistence.*;
 import java.time.LocalDateTime;
 
 
@@ -17,8 +13,8 @@ public class OrderItemDto {
     private Long orderId;
     private Long productId;
     private String productTitle;
-    private Long price;
-    private Long quantity;
+    private int price;
+    private int quantity;
     private LocalDateTime createdAt;
     private LocalDateTime updatedAt;
